Update graphviz dependency
There is a plan to package pygraphviz, which is a separate thing from
the python-graphviz which is part of graphviz. The python subpackages
of graphviz were renemaed back to graphviz-python[23].

Update the Requires to use an unambiguous tag that referes directly to
the module name for python2.

For python3, the dependency on the python2 module doesn't make sense.
Drop it.